Refactor i2cReceive to rpi.c
